docker/convpara1/app/requirements.txt
Flask==0.11.1
Flask-RESTful==0.3.5
Werkzeug==0.10.4
jsonpickle==0.9.4
gunicorn==19.7.1

# From convpara/topical_chat/analysis/requirements.txt
pytorch-ignite
# transformers==2.3.0
transformers==4.3.3
tensorboardX
grpcio==1.24.3
tensorboard==1.13.1

# Fixing versions of dependcies
sentencepiece==0.1.86
